requirements.txt
astroid==3.1.0
blinker==1.8.2
click==8.1.7
dill==0.3.8
Flask==3.0.3
Flask-Cors==3.0.10
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
platformdirs==4.2.1
pylint==3.1.0
python-decouple==3.8
six==1.16.0
tomli==2.0.1
tomlkit==0.12.5
typing_extensions==4.11.0
Werkzeug==3.0.3
